requirements.txt
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
demjson3==3.0.6
fonttools==4.61.1
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
jtutils==0.0.8
kiwisolver==1.4.9
Levenshtein==0.27.3
lxml==6.0.2
matplotlib==3.10.8
mypy_extensions==1.1.0
numpy==2.4.2
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
psycopg2-binary==2.9.11
pyparsing==3.3.2
PySocks==1.7.1
python-csv==0.0.14
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
RapidFuzz==3.14.3
requests==2.32.5
selenium==4.40.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
xlrd==2.0.2
xmltodict==1.0.2
zipp==3.23.0
